refactor: make clean design for navbar and showcase
